other/3do/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
constantly==15.1.0
cryptography==36.0.1
cycler==0.11.0
Deprecated==1.2.13
fonttools==4.28.5
freetype-py==2.2.0
frozenlist==1.2.0
future==0.18.2
google-pasta==0.2.0
h5py==2.10.0
hyperlink==21.0.0
idna==3.3
imageio==2.13.4
incremental==21.3.0
joblib==1.1.0
Keras==2.3.1
Keras-Applications==1.0.8
kiwisolver==1.3.2
matplotlib==3.5.1
multidict==5.2.0
networkx==2.6.3
numpy==1.21.5
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pycparser==2.21
pyglet==1.5.21
PyOpenGL==3.1.0
pyparsing==3.0.6
pyrender==0.1.45
python-dateutil==2.8.2
PyWavelets==1.2.0
PyYAML==6.0
scikit-image==0.19.1
scikit-learn==0.23.2
scipy==1.5.2
tensorboard==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.5.0
tqdm==4.62.3
trimesh==3.9.36
Twisted==21.7.0
txaio==21.2.1
vedo==2021.0.7
vtk==9.0.3
wslink==1.3.0
yarl==1.7.2
zope.interface==5.4.0